Add MCR VSA class
Add a new VSA class for Modular Composite Represnetations (MCR) as
proposed in the paper "Modular Composite Representation" (2014). The MCR
class is a integer quantized alternative to FHRR. This commit introduces
a draft implementation that may change in the future. By now, binding is
implemented as element-wise phase addition and bundle as complex sum.
The MCR class also defines a alternative bundle by majority, which is
not implemented at the moment.

A drawback of this draft implementation is the requirement of passing a
"mod" parameter in function calls. Thus, many function signatures need
to modified to allow passing it by kwargs.