Add **kwargs to allow MCR's mod parameter
Add **kwargs parameter to support the use of mod in the MCR class. This
may not be the best API design to support MCR since it changes a lot of
function signatures. An alternative would be to a MCR class have its own
mod attribute, but early experiments on it were hard to implement in
pytorch.